Filter out premium properties in the members directory
