+++ components/members/components/MemberDirectoryTableView.tsx
@@ -23,8 +23,10 @@ import { DiscordSocialIcon } from 'components/u/components/UserDetails/DiscordSo
 import type { Social } from 'components/u/interfaces';
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { useDateFormatter } from 'hooks/useDateFormatter';
+import { useIsFreeSpace } from 'hooks/useIsFreeSpace';
 import { useMemberProperties } from 'hooks/useMemberProperties';
 import { useUser } from 'hooks/useUser';
+import { PREMIUM_MEMBER_PROPERTIES } from 'lib/members/constants';
 import type { Member } from 'lib/members/interfaces';
 
 import { MemberPropertyTextMultiline } from './MemberDirectoryProperties/MemberPropertyTextMultiline';
@@ -44,8 +46,8 @@ function MemberDirectoryTableRow({ member }: { member: Member }) {
   const discordUsername = (member.profile?.social as Social)?.discordUsername;
   const { space: currentSpace } = useCurrentSpace();
   const { user } = useUser();
-  const { properties = [] } = useMemberProperties();
-  const visibleProperties = properties.filter((property) => property.enabledViews.includes('table'));
+  const { getDisplayProperties } = useMemberProperties();
+  const visibleProperties = getDisplayProperties('table');
   const { showUserProfile } = useUserProfile();
   const { formatDate } = useDateFormatter();
 
@@ -242,9 +244,9 @@ function MemberDirectoryTableRow({ member }: { member: Member }) {
 }
 
 export function MemberDirectoryTableView({ members }: { members: Member[] }) {
-  const { properties = [] } = useMemberProperties();
+  const { getDisplayProperties } = useMemberProperties();
 
-  const visibleProperties = properties.filter((property) => property.enabledViews.includes('table'));
+  const visibleProperties = getDisplayProperties('table');
   return (
     <Table
       size='small'
